+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.0] - 2018-03-22
+
+### Added
+
+- Create links automatically to the issues
+
+### Fixed
+
+- Use UTC date functions to fix decrement issue - [#1](https://github.com/oscarotero/keep-a-changelog/issues/1)
+
 ## [0.3.1] - 2017-12-11
 
 ### Added
@@ -43,6 +53,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 First version
 
+[0.4.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/oscarotero/keep-a-changelog/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.1.0...v0.2.0